requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
cross-web==0.4.1
distro==1.9.0
fastapi==0.115.0
graphql-core==3.2.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
openai==1.54.3
packaging==25.0
pluggy==1.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
strawberry-graphql==0.289.0
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
